customSort is implemented in Array.prototype
