+++ src/index.tsx
@@ -1,5 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
+import '../src/utils/customSort';
 
 import App from './App';
